Only do something if drive letters are defined.
If there are no drive letters defined (e.g. on Linux) then the regular
expression in the connectTo() function will also match links beginning
with ':', '/' or '\'.
Add a if condition that prevents registering a pattern for this plugin
if no drive letter is defined. This should prevent unnecessary load
created by this plugin, when it is not supposed to do anything.

Also this commit adds three tests for the aforementioned cases.